+++ goshipdone.go
@@ -1,0 +1,3 @@
+// Package goshipdone provides an extendable, build-to-publish pipeline for multiple
+// OS-Architecture targets for go packages, built for magefile.
+package goshipdone