+++ internal/modules/checksum.go
@@ -7,8 +7,8 @@ import (
 	"os"
 	"path"
 
-	"github.com/julian7/magelib/ctx"
-	"github.com/julian7/magelib/modules"
+	"github.com/julian7/goshipdone/ctx"
+	"github.com/julian7/goshipdone/modules"
 )
 
 // Checksum calculates checksums of artifacts, and stores them in a checksum file