+++ internal/modules/git_tag.go
@@ -3,8 +3,8 @@ package modules
 import (
 	"fmt"
 
-	"github.com/julian7/magelib/ctx"
-	"github.com/julian7/magelib/modules"
+	"github.com/julian7/goshipdone/ctx"
+	"github.com/julian7/goshipdone/modules"
 	"github.com/magefile/mage/sh"
 )