+++ internal/modules/scp.go
@@ -1,8 +1,8 @@
 package modules
 
 import (
-	"github.com/julian7/magelib/ctx"
-	"github.com/julian7/magelib/modules"
+	"github.com/julian7/goshipdone/ctx"
+	"github.com/julian7/goshipdone/modules"
 	"github.com/magefile/mage/sh"
 )