+++ magelib.go
@@ -1,3 +1,0 @@
-// Package magelib provides an extendable, build-to-publish pipeline for multiple
-// OS-Architecture targets for go packages, built for magefile.
-package magelib